Update game
I'm a Python 2 user, so I hope you'll be able to run this. In this program I basically give the user 5 chances to guess a number from 1 to 10 (by changing the random_number variable you can edit the range of the random number). I suppose that's pretty close to what you were looking for?